pkg/endpoint: refactor identity-related code into own file
Move identity-related code for endpoint into identity.go

Signed-off by: Ian Vernon <ian@cilium.io>
